Added test for checking if internet connection is available
